test(robot): consolidate env unit tests; drop soft redundancy
Merge bridge/claims/gym/spawn suites into one environment/tests/test_robot.py,
rename DatasetWriter tests to test_robot_dataset_writer.py, and drop the noop
release_claim and positive-only tick-loop cases.
